scripts/requirements-compute.txt
numpy>=2.0
pandas>=2.2
scipy>=1.13
scikit-learn>=1.5
psycopg2-binary>=2.9
